requirements.txt
absl-py==1.3.0
accelerate==0.29.3
aiofiles==24.1.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==4.0.0
anyio==4.4.0
appnope==0.1.3
asttokens==2.1.0
async-timeout==4.0.3
attrs==23.2.0
autograd==1.6.2
autograd-gamma==0.5.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.6.2
boto3==1.26.17
botocore==1.29.17
brotlipy==0.7.0
cachetools==5.2.0
certifi==2021.5.30
cffi==1.16.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==7.0.0
contourpy==1.0.6
cycler==0.11.0
dataclasses-json==0.5.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
docopt==0.6.2
entrypoints==0.4
etelemetry==0.3.1
exceptiongroup==1.2.1
executing==1.2.0
faiss-cpu==1.7.4
fastjsonschema==2.19.1
filelock==3.12.2
fire==0.4.0
fitz==0.0.1.dev2
fonttools==4.38.0
formulaic==1.0.1
frontend==0.0.3
frozenlist==1.4.1
fsspec==2024.3.0
future==1.0.0
futures
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.14.1
google-auth-oauthlib==0.4.6
greenlet==3.0.3
grpcio==1.50.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.24.0
idna==3.7
importlib-metadata==5.1.0
InstructorEmbedding==1.0.1
interface-meta==1.3.0
ipython==8.12.3
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==8.6.1
jupyter-core==5.7.2
jupyterlab-pygments==0.3.0
kiwisolver==1.4.4
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-text-splitters==0.2.2
langsmith==0.1.93
lifelines==0.28.0
loguru==0.7.2
looseversion==1.3.0
lxml==5.2.2
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==0.4.3
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
numexpr==2.9.0
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.6
openapi-schema-pydantic==1.2.4
opencv-python==4.6.0.66
orjson==3.10.3
packaging==23.2
pandas==1.5.2
pandocfilters==1.5.1
parso==0.8.3
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.5.0
platformdirs==3.9.1
prompt-toolkit==3.0.33
protobuf==3.20.3
prov==2.0.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.5
pycosat==0.6.3
pycparser==2.22
pydantic==1.10.14
pydeck==0.8.1b0
pydot==2.0.0
Pygments==2.13.0
Pympler==1.0.1
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.0.9
pypdf2==3.0.1
pyspin==1.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.6
pyxnat==1.6.2
PyYAML==6.0
pyzmq==24.0.1
rdflib==6.3.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rouge-score==0.1.2
rpds-py==0.18.0
rsa==4.9
s3transfer==0.6.0
safetensors==0.4.2
scikit-learn==1.0.2
scipy==1.9.3
seaborn==0.12.1
semver==3.0.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
Shapely==1.8.5.post1
shellingham==1.5.4
simplejpeg==1.6.5
simplejson==3.19.2
six==1.16.0
sklearn==0.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
stack-data==0.6.2
starlette==0.37.2
streamlit==1.36.0
streamlit-feedback==0.1.3
StrEnum==0.4.8
sympy==1.12
tenacity==8.2.3
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.1
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.64.1
traitlets==5.5.0
traits==6.3.2
transformers==4.42.4
trubrics==1.3.6
typeguard==2.13.3
typer==0.12.3
typing-extensions==4.10.0
typing-inspect==0.8.0
tzlocal==5.2
urllib3==1.26.19
uvicorn==0.30.1
validators==0.22.0
virtualenv==20.24.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.16.0
yarg==0.1.9
yarl==1.9.4
zipp==3.11.0